+++ pyproject.toml
@@ -12,6 +12,8 @@ beautifulsoup4 = "^4.12.3"
 pandas = "^2.2.3"
 pydantic = "^2.10.6"
 pydantic-settings = "^2.7.1"
+black = "^24.10.0"
+isort = "^5.13.2"
 
 
 [build-system]
