+++ rent_radar/analysis.py
@@ -1,0 +1,16 @@
+from typing import Any
+
+
+def find_flagged_listings() -> list[dict[str, Any]]:
+    """
+    Identify any listings with a price change after 1/5/2025 that is >= 10% from previous price.
+    We'll query the price_history table and compare consecutive entries.
+    """
+    raise NotImplementedError
+
+
+def export_flagged_to_csv(flagged_listings: list[dict[str, Any]]):
+    """
+    Writes flagged listings to CSV.
+    """
+    raise NotImplementedError
